pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "tracevault"
version = "0.1.0"
description = "Tamper-proof audit trails for LangGraph/LangChain agents with HMAC-chained entries and LLM-generated summaries"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.30",
    "openai>=1.12.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
]

[project.scripts]
tracevault = "tracevault.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["tracevault*"]
